Remove Test Mail and Minor Fix
1. Only send notification when host is end with ust.hk
2.  Minor display issue fix
3. ProgramType image file removal only when exist

+++ Controllers/ProgramTypeController.cs
@@ -124,8 +124,11 @@ namespace SchoolOfScience.Controllers
                     }
                     try
                     {
-                        System.IO.File.Move(sourceFilepath, destFilepath);
-                        programtype.image_filepath = "~/Images/ProgramType/" + programtype.id.ToString();
+                        if (!System.IO.File.Exists(destFilepath))
+                        {
+                            System.IO.File.Move(sourceFilepath, destFilepath);
+                            programtype.image_filepath = "~/Images/ProgramType/" + programtype.id.ToString();
+                        }
                     }
                     catch (Exception e)
                     {
